Move write tests to a separate file
